+++ fluXis/Screens/Select/Search/Dropdown/SearchDropdownKeymode.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using fluXis.Audio;
 using fluXis.Graphics.Sprites.Text;
+using fluXis.Graphics.UserInterface.Buttons;
 using fluXis.Graphics.UserInterface.Color;
 using fluXis.Graphics.UserInterface.Interaction;
 using fluXis.Utils;
